requirements.txt
requests
python-dotenv
mysql-connector-python
fastapi
uvicorn
slowapi>=0.1.9
